+++ src/components/navbar/navbar.tsx
@@ -14,10 +14,10 @@ interface Props {
 function Navbar(props: Props) {
     return (
         <React.Fragment>
-            <div id='RegularNavbar' className='z-30 hidden lg:block'>
+            <div className='z-30 hidden lg:block'>
                 <VisualRegularNavbar logOut={props.logOut} />
             </div>
-            <div id='MobileNavbar' className='z-30 block lg:hidden'>
+            <div className='z-30 block lg:hidden'>
                 <VisualMobileNavbar
                     logOut={props.logOut}
                     navbarState={props.navbarState}
