+++ src/pages/pages-types/dashboard-page.tsx
@@ -3,9 +3,6 @@ import {connect} from 'react-redux';
 import {types} from '@types';
 import {Navbar} from '@components';
 
-// TODO: Avoid plain css
-import '@styles/dashboard-page.css';
-
 function DashBoardPage(props: {
     children: React.ReactNode;
     navbarState: types.NavbarState;
