--- src/styles/dashboard-page.css
@@ -1,24 +1,0 @@
-#RegularContent {
-    /*position: relative;*/
-    /*margin-left: 16rem;*/
-    /*width: calc(100vw - 16rem);*/
-    /*min-height: calc(100vh);*/
-}
-
-#MobileContent {
-    position: relative;
-    width: calc(100vw - 4rem);
-    padding: 0.8rem 2rem 2rem 2rem;
-    min-height: calc(100% - 4rem);
-}
-
-#MobileContent h3:first-child {
-    padding-left: 2.5rem;
-}
-
-#ConfigEditor,
-#ControlStrip {
-    margin-left: 20rem;
-    width: calc(100vw - 36rem);
-}
-
